TripService: change type of variables - driverId and carId was changed from long to int

+++ backend/src/main/java/com/hardelele/ct/services/TripService.java
@@ -9,8 +9,8 @@ public interface TripService {
 
     List<TripEntity> getAll();
     Optional<TripEntity> getOneById(Long id);
-    void add(String departure, String destination, long driverId, long carId, double driverPrice, String startDate, String endDate);
-    void update(long id, String departure, String destination, long driverId, long carId, double driverPrice, String startDate, String endDate);
+    void add(String departure, String destination, int driverId, int carId, double driverPrice, String startDate, String endDate);
+    void update(long id, String departure, String destination, int driverId, int carId, double driverPrice, String startDate, String endDate);
     void delete(long id);
     void deleteAll();
 }
